web_demo/requirements.txt
# Backend
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.0
python-multipart==0.0.9
jinja2==3.1.4
aiofiles==24.1.0
python-dotenv==1.0.1

# ML / Image
torch
torchvision
pillow==10.4.0
numpy==1.26.4

# Optional tooling (can be removed if not needed)
# sseclient-py==1.8.0

